Tease out the connector structure even more (inbound, backend, outbound)
