+++ nexus/ui_connector/router/workflow.go
@@ -1,23 +1,26 @@
+// Package router implements RouterWorkflow, the single Temporal workflow type shared by
+// every inbound platform (Slack, Teams, ...). It defines the two ports platform code
+// plugs into - OutboundDriver (delivers a response back to the user) and BackendDriver
+// (produces the response) - plus the shared Input envelope inbound drivers construct to
+// start it. Platform packages (slack/, teams/, agent/) import router to implement these
+// ports; router never imports platform code.
 package router
 
 import (
 	"fmt"
 
-	"github.com/temporal-community/temporal-agent-harness/nexus/ui_connector/inbound"
-	"github.com/temporal-community/temporal-agent-harness/nexus/ui_connector/outbound"
-	"github.com/temporal-community/temporal-agent-harness/nexus/ui_connector/wire"
 	"go.temporal.io/sdk/workflow"
 )
 
 // WorkflowName is the single registered name for RouterWorkflow.Run.
 const WorkflowName = "RouterWorkflow"
 
-// RouterWorkflow routes a single interaction between an inbound driver and an outbound driver.
+// RouterWorkflow routes a single interaction between an outbound driver and a backend.
 // It is the single entry point for all connector interactions, and is responsible for
-// forwarding input to the outbound driver and reacting to the shape of the result.
+// forwarding input to the backend and reacting to the shape of the result.
 type RouterWorkflow struct {
-	inbound  inbound.Driver
-	outbound outbound.Driver
+	outbound OutboundDriver
+	backend  BackendDriver
 }
 
 // RouterWorkflowID returns a stable, unique workflow ID for an interaction.
@@ -27,20 +30,21 @@ func RouterWorkflowID(identity, sessionID, interactionID string) string {
 	return fmt.Sprintf("connector-%s-%s-%s", identity, sessionID, interactionID)
 }
 
-func NewRouterWorkflow(inboundDriver inbound.Driver, outboundDriver outbound.Driver) *RouterWorkflow {
-	return &RouterWorkflow{inbound: inboundDriver, outbound: outboundDriver}
+func NewRouterWorkflow(outboundDriver OutboundDriver, backendDriver BackendDriver) *RouterWorkflow {
+	return &RouterWorkflow{outbound: outboundDriver, backend: backendDriver}
 }
 
-// Run is the workflow that makes a single turn between inbound and outbound durable.
-func (w *RouterWorkflow) Run(ctx workflow.Context, input wire.Input) error {
-	result, err := w.outbound.StartTurn(ctx, input)
+// Run is the workflow that makes a single turn between the outbound driver and the
+// backend durable.
+func (w *RouterWorkflow) Run(ctx workflow.Context, input Input) error {
+	result, err := w.backend.StartTurn(ctx, input)
 	if err != nil {
 		workflow.GetLogger(ctx).Warn("RouterWorkflow: StartTurn failed", "error", err)
 		return nil
 	}
 
 	if approval := input.Approval; approval != nil {
-		if err := w.inbound.AcknowledgeApproval(ctx, inbound.ApprovalAcknowledgementInput{
+		if err := w.outbound.AcknowledgeApproval(ctx, ApprovalAcknowledgementInput{
 			TextMetadata: textMetadata(input, ""),
 			PromptID:     approval.ActivityID,
 			ToolName:     approval.ToolName,
@@ -54,15 +58,15 @@ func (w *RouterWorkflow) Run(ctx workflow.Context, input wire.Input) error {
 	switch {
 	case result.Reply != "":
 		// An immediate, synchronous answer - no turn was created, nothing to poll.
-		return w.inbound.PostMessage(ctx, textMetadata(input, result.Reply))
+		return w.outbound.PostMessage(ctx, textMetadata(input, result.Reply))
 
-	case result.Handle != nil && !w.inbound.SupportsStreaming(input):
-		// Collect the complete response when the inbound conversation does not
+	case result.Handle != nil && !w.outbound.SupportsStreaming(input):
+		// Collect the complete response when the outbound conversation does not
 		// support native streaming, then post it as a single message.
 		return w.postResp(ctx, *result.Handle, input)
 
 	case result.Handle != nil:
-		// A turn was created; consume its response stream and deliver it inbound.
+		// A turn was created; consume its response stream and deliver it outbound.
 		return w.streamResp(ctx, *result.Handle, input)
 
 	default:
@@ -71,8 +75,8 @@ func (w *RouterWorkflow) Run(ctx workflow.Context, input wire.Input) error {
 	}
 }
 
-func textMetadata(input wire.Input, text string) inbound.TextMetadata {
-	metadata := inbound.TextMetadata{
+func textMetadata(input Input, text string) TextMetadata {
+	metadata := TextMetadata{
 		SessionID: input.SessionID,
 		ThreadID:  input.ThreadID(),
 		SenderID:  input.SenderID(),
@@ -89,13 +93,13 @@ func textMetadata(input wire.Input, text string) inbound.TextMetadata {
 }
 
 // postResp polls a turn to completion and posts all text as one message for
-// inbound conversations that do not support native streams.
-func (w *RouterWorkflow) postResp(ctx workflow.Context, handle outbound.TurnHandle, input wire.Input) error {
+// outbound conversations that do not support native streams.
+func (w *RouterWorkflow) postResp(ctx workflow.Context, handle TurnHandle, input Input) error {
 	cursor := handle.StreamHeadOffset
 	fullText := ""
 
 	for {
-		res, err := w.outbound.PollTurn(ctx, handle, cursor)
+		res, err := w.backend.PollTurn(ctx, handle, cursor)
 		if err != nil {
 			workflow.GetLogger(ctx).Warn("postResp: PollTurn failed", "error", err)
 			return nil
@@ -106,13 +110,13 @@ func (w *RouterWorkflow) postResp(ctx workflow.Context, handle outbound.TurnHand
 			if fullText == "" {
 				return nil
 			}
-			return w.inbound.PostMessage(ctx, textMetadata(input, fullText))
+			return w.outbound.PostMessage(ctx, textMetadata(input, fullText))
 		}
 
 		for _, delta := range res.Deltas {
 			if delta.ApprovalRequested != nil {
 				req := delta.ApprovalRequested
-				if err := w.inbound.PostApprovalPrompt(ctx, inbound.ApprovalPromptInput{
+				if err := w.outbound.PostApprovalPrompt(ctx, ApprovalPromptInput{
 					TextMetadata: textMetadata(input, ""),
 					ToolID:       req.ToolID,
 					ToolName:     req.ToolName,
@@ -128,27 +132,30 @@ func (w *RouterWorkflow) postResp(ctx workflow.Context, handle outbound.TurnHand
 				if fullText == "" {
 					return nil
 				}
-				return w.inbound.PostMessage(ctx, textMetadata(input, fullText))
+				return w.outbound.PostMessage(ctx, textMetadata(input, fullText))
 			}
 		}
 	}
 }
 
-// streamResp polls the outbound driver for a started turn and streams each delta back
-// through the inbound driver, until the turn closes. This loop is generic over any
-// outbound/inbound pairing: it only deals in outbound.Delta and inbound.Driver calls.
-func (w *RouterWorkflow) streamResp(ctx workflow.Context, handle outbound.TurnHandle, input wire.Input) error {
+// streamResp polls the backend for a started turn and streams each delta back
+// through the outbound driver, until the turn closes. This loop is generic over any
+// backend/outbound pairing: it only deals in Delta and OutboundDriver calls.
+func (w *RouterWorkflow) streamResp(ctx workflow.Context, handle TurnHandle, input Input) error {
 	cursor := handle.StreamHeadOffset
 
-	// Start the initial inbound stream before polling.
+	// Start the initial outbound stream before polling. If the outbound driver can't
+	// open a stream (e.g. its activity worker died mid-attempt with no retries left),
+	// fall back to collecting the full response and posting it as one message rather
+	// than silently dropping it.
 	streamHandle, err := w.beginStream(ctx, input)
 	if err != nil {
-		workflow.GetLogger(ctx).Error("streamResp: stream begin failed", "error", err)
-		return nil
+		workflow.GetLogger(ctx).Warn("streamResp: stream begin failed, falling back to postResp", "error", err)
+		return w.postResp(ctx, handle, input)
 	}
 
 	for {
-		res, err := w.outbound.PollTurn(ctx, handle, cursor)
+		res, err := w.backend.PollTurn(ctx, handle, cursor)
 		if err != nil {
 			workflow.GetLogger(ctx).Warn("streamResp: PollTurn failed", "error", err)
 			w.endStream(ctx, input, streamHandle)
@@ -165,14 +172,14 @@ func (w *RouterWorkflow) streamResp(ctx workflow.Context, handle outbound.TurnHa
 		for _, delta := range res.Deltas {
 			if delta.ApprovalRequested != nil {
 				req := delta.ApprovalRequested
-				// Some inbound drivers must finish the current stream before posting an
+				// Some outbound drivers must finish the current stream before posting an
 				// approval card so the messages appear in order. Clearing the handle makes
 				// the next text delta start a new stream.
 				if streamHandle != nil && streamHandle.CloseBeforeApproval {
 					w.endStream(ctx, input, streamHandle)
 					streamHandle = nil
 				}
-				if err := w.inbound.PostApprovalPrompt(ctx, inbound.ApprovalPromptInput{
+				if err := w.outbound.PostApprovalPrompt(ctx, ApprovalPromptInput{
 					TextMetadata: textMetadata(input, ""),
 					ToolID:       req.ToolID,
 					ToolName:     req.ToolName,
@@ -204,12 +211,12 @@ func (w *RouterWorkflow) streamResp(ctx workflow.Context, handle outbound.TurnHa
 	}
 }
 
-func (w *RouterWorkflow) beginStream(ctx workflow.Context, input wire.Input) (*inbound.StreamHandle, error) {
+func (w *RouterWorkflow) beginStream(ctx workflow.Context, input Input) (*StreamHandle, error) {
 	conversationType := ""
 	if input.Message != nil {
 		conversationType = input.Message.ConversationType
 	}
-	streamHandle, err := w.inbound.BeginStream(ctx, inbound.BeginStreamInput{
+	streamHandle, err := w.outbound.BeginStream(ctx, BeginStreamInput{
 		TextMetadata:     textMetadata(input, ""),
 		ConversationType: conversationType,
 	})
@@ -221,14 +228,14 @@ func (w *RouterWorkflow) beginStream(ctx workflow.Context, input wire.Input) (*i
 
 func (w *RouterWorkflow) updateStream(
 	ctx workflow.Context,
-	input wire.Input,
-	handle *inbound.StreamHandle,
+	input Input,
+	handle *StreamHandle,
 	delta string,
 ) {
 	if handle == nil {
 		return
 	}
-	if err := w.inbound.UpdateStream(ctx, inbound.UpdateStreamInput{
+	if err := w.outbound.UpdateStream(ctx, UpdateStreamInput{
 		TextMetadata: textMetadata(input, ""),
 		Handle:       *handle,
 		Delta:        delta,
@@ -237,11 +244,11 @@ func (w *RouterWorkflow) updateStream(
 	}
 }
 
-func (w *RouterWorkflow) endStream(ctx workflow.Context, input wire.Input, handle *inbound.StreamHandle) {
+func (w *RouterWorkflow) endStream(ctx workflow.Context, input Input, handle *StreamHandle) {
 	if handle == nil {
 		return
 	}
-	if err := w.inbound.FinishStream(ctx, inbound.FinishStreamInput{
+	if err := w.outbound.FinishStream(ctx, FinishStreamInput{
 		TextMetadata: textMetadata(input, ""),
 		Handle:       *handle,
 	}); err != nil {
