+++ nexus/ui_connector/slack/slack_app_manifest.yaml
